Update remote_cwd.sh
I haven't tested this live. But I believe it should work fine!

On the macOS end you will need to update the fact this this script can now return "error" and if so, we should abstain from sending ssh events to autocomplete as we don't know the working directory. We should also signal this in the debugger